requirements.txt
gc
kagglehub
matplotlib.pyplot
nltk
nltk.corpus
nltk.tokenize
numpy
pandas
re
seaborn
sklearn.feature_extraction.text
sklearn.metrics
sklearn.model_selection
sklearn.preprocessing
sklearn.utils
tensorflow.keras.layers
tensorflow.keras.models
tensorflow.keras.preprocessing.sequence
tensorflow.keras.preprocessing.text
tqdm
streamlit
